pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "LogoBgVanisher"  # Your package name
version = "0.1.0"
description = "A package to remove (or make transparent) the background from AI-generated logos."
authors = [{ name = "Dragos Josan", email = "dragosjosan@gmail.com" }]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "Pillow",
    "rembg",
    "pytest"
]

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
background_logo_remover = "logobgvanisher.main:main"